Added: active record deletion marker
git-svn-id: svn://svn.orena.org/orena/trunk@131 d1990dbc-1c88-9740-88cc-0edcd7010ae0

+++ O/src/Dao/ActiveRecord.phps
@@ -33,7 +33,7 @@ abstract class O_Dao_ActiveRecord implements ArrayAccess {
 	 * @var Array
 	 */
 	private static $injected_methods = Array ();
-	
+
 	/**
 	 * Array of SQL field values of a database record
 	 *
@@ -46,6 +46,12 @@ abstract class O_Dao_ActiveRecord implements ArrayAccess {
 	 * @var array
 	 */
 	private $changed = array ();
+	/**
+	 * Deletion process marker
+	 *
+	 * @var int
+	 */
+	private $__is_deleted = 0;
 
 	/**
 	 * Creates a new object in database
@@ -74,13 +80,13 @@ abstract class O_Dao_ActiveRecord implements ArrayAccess {
 				throw $e;
 			}
 		}
-		
+
 		$this->reload();
 		$class = get_class( $this );
 		self::$objs[ $class ][ $this->fields[ "id" ] ] = $this;
-		
+
 		if (O_Dao_TableInfo::get( $class )->getParam( "signal" )) {
-			O_Dao_Signals::fire( O_Dao_Signals::EVENT_CREATE, O_Dao_TableInfo::get( $class )->getParam( "signal" ), 
+			O_Dao_Signals::fire( O_Dao_Signals::EVENT_CREATE, O_Dao_TableInfo::get( $class )->getParam( "signal" ),
 					$class, $this, $this->fields[ "id" ] );
 		}
 	}
@@ -97,12 +103,12 @@ abstract class O_Dao_ActiveRecord implements ArrayAccess {
 			return $this->fields[ "id" ];
 		if (strpos( $name, "." )) {
 			list ($name, $subreq) = explode( ".", $name, 2 );
-			return $this->getFieldInfo( $name )->getMappedQuery( $this, 
+			return $this->getFieldInfo( $name )->getMappedQuery( $this,
 					isset( $this->fields[ $name ] ) ? $this->fields[ $name ] : null, $subreq );
 		}
-		
-		return $this->getFieldInfo( $name )->getValue( $this, 
-				isset( $this->fields[ $name ] ) ? $this->fields[ $name ] : null, 
+
+		return $this->getFieldInfo( $name )->getValue( $this,
+				isset( $this->fields[ $name ] ) ? $this->fields[ $name ] : null,
 				array_key_exists( $name, $this->fields ) );
 	}
 
@@ -183,7 +189,7 @@ abstract class O_Dao_ActiveRecord implements ArrayAccess {
 		$fieldInfo = O_Dao_TableInfo::get( get_class( $this ) )->getFieldInfo( $name );
 		if (!$fieldInfo)
 			throw new O_Ex_NotFound( "Unknown field: $name." );
-		
+
 		return $fieldInfo;
 	}
 
@@ -196,21 +202,21 @@ abstract class O_Dao_ActiveRecord implements ArrayAccess {
 	{
 		if (!count( $this->changed ))
 			return true;
-		
+
 		$query = new O_Dao_Query( $this );
 		$query->test( "id", $this->fields[ "id" ] );
-		
+
 		foreach ($this->changed as $name => $value) {
 			$query->field( $name, $value );
 		}
-		
+
 		if ($query->update()) {
 			foreach ($this->changed as $name => $value)
 				$this->fields[ $name ] = $value;
 			$this->changed = array ();
 			return true;
 		}
-		
+
 		return false;
 	}
 
@@ -242,7 +248,7 @@ abstract class O_Dao_ActiveRecord implements ArrayAccess {
 	{
 		if (isset( self::$objs[ $class ][ $id ] ))
 			return self::$objs[ $class ][ $id ];
-		
+
 		self::$objs[ $class ][ $id ] = unserialize( sprintf( 'O:%d:"%s":0:{}', strlen( $class ), $class ) );
 		if (!isset( $row[ "id" ] )) {
 			$query = new O_Dao_Query( $class );
@@ -280,16 +286,19 @@ abstract class O_Dao_ActiveRecord implements ArrayAccess {
 	 */
 	public function delete()
 	{
+		if ($this->__is_deleted)
+			return;
+		$this->__is_deleted = 1;
 		$fields = O_Dao_TableInfo::get( $this )->getFields();
 		foreach ($fields as $name => $field)
 			$field->deleteThis( $this, isset( $this->fields[ $name ] ) ? $this->fields[ $name ] : null );
-		
+
 		$query = new O_Dao_Query( $this );
 		$query->test( "id", $this->fields[ "id" ] )->delete();
-		
+
 		$class = get_class( $this );
 		if (O_Dao_TableInfo::get( $class )->getParam( "signal" )) {
-			O_Dao_Signals::fire( O_Dao_Signals::EVENT_DELETE, O_Dao_TableInfo::get( $class )->getParam( "signal" ), 
+			O_Dao_Signals::fire( O_Dao_Signals::EVENT_DELETE, O_Dao_TableInfo::get( $class )->getParam( "signal" ),
 					$class, $this, $this->fields[ "id" ] );
 		}
 	}
@@ -390,7 +399,7 @@ abstract class O_Dao_ActiveRecord implements ArrayAccess {
 	public function getParam( $paramName, $fieldName = null, $parseAsArray = 0 )
 	{
 		$tableInfo = O_Dao_TableInfo::get( get_class( $this ) );
-		return $fieldName ? $tableInfo->getFieldInfo( $fieldName )->getParam( $paramName, $parseAsArray ) : $tableInfo->getParam( 
+		return $fieldName ? $tableInfo->getFieldInfo( $fieldName )->getParam( $paramName, $parseAsArray ) : $tableInfo->getParam(
 				$fieldName, $parseAsArray );
 	}
